File

+++ controllers/admin.js
@@ -2,6 +2,8 @@ const Product = require('../models/product');
 
 const mongoose = require('mongoose');
 
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
 exports.getAddProduct = (req, res, next) => {
@@ -23,18 +25,33 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
- // const _id = new mongoose.Types.ObjectId('5c298afc3c080c314829b241');
   const title = req.body.title;
   const price = req.body.price;
   const description = req.body.description;
-  const imageURL = req.body.imageURL;
+  const image = req.file;
   const userId = req.user;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/add-product',
+      editMode: false,
+      isAuthenticated: req.session.isLoggedIn,
+      editMode: false,
+      hasError: true,
+      errorMessage: 'Attached file is not an image!',
+      validationResult: [],
+      product: {
+        title: title,
+        price: price,
+        description: description
+      }
+    });
+  }
 
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
-    console.log(errors.array());
-
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/add-product',
@@ -47,14 +64,15 @@ exports.postAddProduct = (req, res, next) => {
       product: {
         title: title,
         price: price,
-        description: description,
-        imageURL: imageURL
+        description: description
       }
     });
   }
 
+  const imageURL = image.path;
+
   const product = new Product({
-   // _id,
+    // _id,
     title,
     price,
     description,
@@ -68,6 +86,7 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/');
     })
     .catch(err => {
+      console.log(err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -104,7 +123,7 @@ exports.postEditProduct = (req, res, next) => {
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
   const updatedDescription = req.body.description;
-  const updatedImageURL = req.body.imageURL;
+  const image = req.file;
   const idProd = req.body.productId;
 
   const errors = validationResult(req);
@@ -113,7 +132,6 @@ exports.postEditProduct = (req, res, next) => {
     return res.status(422).render('admin/edit-product', {
       product: {
         title: updatedTitle,
-        imageURL: updatedImageURL,
         price: updatedPrice,
         description: updatedDescription,
         _id: idProd
@@ -132,7 +150,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageURL = updatedImageURL;
+      if (image) {
+        fileHelper.deleteFile(product.imageURL);
+        product.imageURL = image.path;
+      }
       return product.save();
     })
     .then(result => {
@@ -166,7 +187,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.getDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndDelete(prodId)
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found!'));
+      }
+      fileHelper.deleteFile(product.imageURL);
+      return Product.findByIdAndDelete(prodId);
+    })
     .then(result => {
       console.log('Product Deleted');
       res.redirect('/admin/products');
